refactor(open-with): split input and overlay modules
